trying to add conditional import to app.py
