+++ Accounts.json
@@ -1,0 +1,6 @@
+{
+  "official_bot": 2585899559,
+  "official_group": 705716007,
+  "testing_bot": 3292297816,
+  "testing_group": 950540652
+}
