+++ dbUserName.json
@@ -1,0 +1,6 @@
+{
+  "username": "qqbot",
+  "password": "qqbotpassword",
+  "ip": "152.136.189.165",
+  "database": "qqbot"
+}
